+++ src/lib.rs
@@ -25,6 +25,7 @@ pub mod graphlog;
 pub mod merge_tools;
 pub mod operation_templater;
 mod progress;
+pub mod template_builder;
 pub mod template_parser;
 pub mod templater;
 pub mod text_util;
